@@ -1,14 +1,6 @@
-/*
- * Angular 2 decorators and services
- */
-import { Component, ViewEncapsulation } from '@angular/core';
+import {Component, ViewEncapsulation} from '@angular/core';
+import {AppState} from './app.service';
 
-import { AppState } from './app.service';
-
-/*
- * App Component
- * Top Level Component
- */
 @Component({
   selector: 'app',
   encapsulation: ViewEncapsulation.None,
@@ -19,19 +11,12 @@ import { AppState } from './app.service';
   templateUrl: './app.component.html'
 })
 export class AppComponent {
-  angularclassLogo = 'assets/img/angularclass-avatar.png';
-  name = 'Angular 2 Webpack Starter';
-  url = 'https://twitter.com/AngularClass';
-
-  constructor(
-    public appState: AppState) {
-
+  constructor(public appState: AppState) {
   }
 
   ngOnInit() {
     console.log('Initial App State', this.appState.state);
   }
-
 }
 
 /*
