@@ -1,48 +1,19 @@
 import {Component} from '@angular/core';
 
 import {UserApi} from '../../shared/sdk';
-import {AppState} from '../../app.service';
-import {Title} from './title';
-import {XLarge} from './x-large';
 
 @Component({
-  // The selector is what angular internally uses
-  // for `document.querySelectorAll(selector)` in our index.html
-  // where, in this case, selector is the string 'home'
   selector: 'home',  // <home></home>
-  // We need to tell Angular's Dependency Injection which providers are in our app.
-  providers: [
-    Title
-  ],
-  // Our list of styles in our component. We may add more to compose many styles together
+  providers: [],
   styleUrls: ['./home.component.scss'],
-  // Every Angular template is first compiled by the browser before Angular runs it's compiler
   templateUrl: './home.component.html'
 })
 export class HomeComponent {
-  // Set our default values
-  localState = {value: ''};
-  // TypeScript public modifiers
-  constructor(public appState: AppState, public title: Title, public userApi: UserApi) {
-
+  constructor(public users: UserApi) {
   }
 
   ngOnInit() {
-    console.log('hello `Home` component');
-    // this.title.getData().subscribe(data => this.data = data);
-    this.userApi.getCurrent()
-      .subscribe((user) => {
-        console.log(user);
-        console.log(this.userApi.isAuthenticated());
-      }, console.error.bind(console));
-
-    this.userApi.ldapLookup('lng25')
+    this.users.ldapLookup('lng25')
       .subscribe(console.log.bind(console), console.error.bind(console));
-  }
-
-  submitState(value: string) {
-    console.log('submitState', value);
-    this.appState.set('value', value);
-    this.localState.value = '';
   }
 }
